Adding python packages backup fallback

+++ shell/python_packages.sh
@@ -1,6 +1,8 @@
 function backup_python_packages() {
     mkdir -p $SYNC_PATH/python-packages/
-    mv $SYNC_PATH/python-packages/$(hostname)_current.txt $SYNC_PATH/python-packages/$(hostname)_$(date +"%Y%m%d%H%M%S").txt
+    if [ -f "$SYNC_PATH/python-packages/$(hostname)_current.txt" ]; then
+        mv $SYNC_PATH/python-packages/$(hostname)_current.txt $SYNC_PATH/python-packages/$(hostname)_$(date +"%Y%m%d%H%M%S").txt
+    fi
     pip freeze | grep '==' > $SYNC_PATH/python-packages/$(hostname)_current.txt.tmp
 }
